requirements.txt
tensorflow-gpu==2.1.0
numpy==1.17.0
opencv-python==4.1.0.25
moviepy